Convert kcal breakdown to float instead of integer

+++ src/tymenu/menu/forms.py
@@ -21,9 +21,9 @@ class RecipeForm(FlaskForm):
         coerce=int,
     )
     kcal = FloatField("Calories:", validators=[Optional(strip_whitespace=True)])
-    protein_gram = IntegerField("Protein (g):", validators=[Optional(strip_whitespace=True)])
-    carb_gram = IntegerField("Carbs (g):", validators=[Optional(strip_whitespace=True)])
-    fat_gram = IntegerField("Fat (g):", validators=[Optional(strip_whitespace=True)])
+    protein_gram = FloatField("Protein (g):", validators=[Optional(strip_whitespace=True)])
+    carb_gram = FloatField("Carbs (g):", validators=[Optional(strip_whitespace=True)])
+    fat_gram = FloatField("Fat (g):", validators=[Optional(strip_whitespace=True)])
     ingredients = PageDownField("Ingredients:", validators=[DataRequired()])
     instructions = PageDownField("Instructions:", validators=[DataRequired()])
     keywords = StringField("Keywords:", validators=[DataRequired()])
